@@ -3,11 +3,10 @@
 // http://www.apache.org/licenses/LICENSE-2.0
 
 const config = require("./jest.config");
+const glob = require('glob');
+const path = require('path');
 
 module.exports = {
   ...config,
-  globals: {
-    ...(config.globals || {}),
-    __INTEGRATION__: true
-  }
+  projects: glob.sync(path.join(__dirname, 'packages/*/jest.integration.config.js')),
 };
