@@ -4,8 +4,8 @@
  * http://www.apache.org/licenses/LICENSE-2.0
  */
 import { Context, SqrlKey } from "sqrl";
-import { RedisInterface, createRedisKey } from "./RedisService";
 import { TOTAL_COUNT_EXPIRY_SEC } from "./RedisCountService";
+import { RedisInterface, createRedisKey } from "./RedisInterface";
 
 export interface RedisBucketCountInterface {
   bump(
