@@ -3,12 +3,12 @@
  * Licensed under the Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-import { LocalFilesystem, Filesystem } from "../api/filesystem";
+import { Filesystem } from "../api/filesystem";
 import { SqrlParserSourceOptions } from "../compile/SqrlParserState";
-import { basename, dirname } from "path";
 import { parseSqrl } from "../parser/SqrlParse";
 import { StatementAst } from "../ast/Ast";
 import invariant from "../jslib/invariant";
+import { LocalFilesystem, splitPath } from "../node/LocalFilesystem";
 
 export function statementsFromString(
   source: string,
@@ -47,9 +47,10 @@ export function sourceOptionsFromPath(
     customFunctions?: Set<string>;
   } = {}
 ): SqrlParserSourceOptions {
+  const {basename,dirname} = splitPath(path);
   return sourceOptionsFromFilesystem(
-    new LocalFilesystem(dirname(path)),
-    basename(path),
+    new LocalFilesystem(dirname),
+    basename,
     options
   );
 }
